move api into controllers so that double colon could be bypass
