--- app/api/api_root.rb
@@ -1,4 +1,0 @@
-class ApiRoot < Grape::API
-  format :json
-  mount V0::Base
-end
