--- app/api/v0/entities/base.rb
@@ -1,8 +1,0 @@
-module V0
-  module Entities
-    class Base < Grape::Entity
-      format_with(:iso8601) { |dt| dt.iso8601 }
-    end
-  end
-
-end
