--- app/api/v0/entities/record.rb
@@ -1,9 +1,0 @@
-module V0
-  module Entities
-    class Record < Entities::Base
-      expose :price
-      expose :price_type
-      expose :expiration
-    end
-  end
-end
