--- app/api/v0/ping.rb
@@ -1,8 +1,0 @@
-module V0
-  class Ping < Grape::API
-    desc "Ping pong"
-    get "/ping" do
-      "pong"
-    end
-  end
-end
