+++ app/controllers/api/api_root.rb
@@ -1,0 +1,4 @@
+class Api::ApiRoot < Grape::API
+  format :json
+  mount Api::V0::Base
+end
