+++ app/controllers/api/v0/base.rb
@@ -1,0 +1,9 @@
+class Api::V0::Base < Grape::API
+  version 'v0', using: :path
+  helpers Api::V0::Helpers
+  mount Ping
+  mount Tutorial
+  mount Registration
+  mount Login
+  mount Purchase
+end
