+++ app/controllers/api/v0/entities/base.rb
@@ -1,0 +1,4 @@
+
+class Api::V0::Entities::Base < Grape::Entity
+  format_with(:iso8601) { |dt| dt.iso8601 }
+end
