+++ app/controllers/api/v0/entities/record.rb
@@ -1,0 +1,7 @@
+
+class Api::V0::Entities::Record < Api::V0::Entities::Base
+  expose :price
+  expose :price_type
+  expose :expiration
+end
+
