+++ app/controllers/api/v0/ping.rb
@@ -1,0 +1,6 @@
+class Api::V0::Ping < Grape::API
+  desc "Ping pong"
+  get "/ping" do
+    "pong"
+  end
+end
