+++ config/application.rb
@@ -10,10 +10,10 @@ module SnapaskExam
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 5.2
-    # config.factory_bot.definition_file_paths = ["spec/factories"]
-    # config.generators do |g|
-    #   g.factory_bot dir: 'spec/factories'
-    # end
+    config.factory_bot.definition_file_paths = ["spec/factories"]
+    config.generators do |g|
+      g.factory_bot dir: 'spec/factories'
+    end
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
