+++ config/routes.rb
@@ -4,6 +4,6 @@ Rails.application.routes.draw do
     registrations: 'user/registrations'
   }
   resources :tutorials
-  mount ApiRoot => '/api'
+  mount Api::ApiRoot => '/api'
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
